+++ src/lib.rs
@@ -20,6 +20,8 @@ pub mod rendering;
 pub mod schema;
 pub mod search;
 pub mod server;
+#[cfg(feature = "ssr")]
+pub mod static_assets;
 pub mod storage;
 #[cfg(test)]
 pub mod test_utils;